+++ lax/__init__.py
@@ -1,1 +1,3 @@
-__version__ = '0.3.0'
+__version__ = '0.4.0'
+
+from . import lichen, lichens